backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi>=0.104.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
logging==0.4.9.6
openai>=1.3.3
packaging==24.2
pluggy==1.5.0
pyactiveresource==2.2.2
pydantic>=2.5.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-env==1.1.5
python-dotenv>=1.0.0
PyYAML>=6.0.1
requests==2.32.3
ShopifyAPI==12.7.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
tqdm==4.67.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn>=0.24.0
